requirements.txt
attrs==19.3.0
cffi==1.14.0
click==7.1.1
coverage==5.0.4
cryptography==2.9
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
six==1.14.0
SQLAlchemy==1.3.16
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
requests
pytest
